test: split security-regression.test.ts into per-domain files (#1188)
- 8 new files, one per security domain (auth, rbac, headers, prompt-guard,
  sockets, stream-tickets, jwt, infra)
- Each file is self-contained (own imports, mocks, beforeAll setup)
- Re-enabled the 11 tests skipped in #1185 / #1110 due to cross-describe
  vi.resetModules() interactions:
  * 10 SSE Stream Ticket cases (#1112) — landed in stream-tickets file
  * 1 PATCH /api/users/:id assertUser fail-loud (#1110) — landed in rbac file
- Original security-regression.test.ts deleted
- All new files pass 100% in isolation AND together (140 tests, 0 skipped)

File breakdown (it() count / total runtime tests):
  auth.test.ts             17 / 17  — auth sweep, rate limit, trust proxy, OIDC
  rbac.test.ts             21 / 21  — remediation, pcap, edge jobs, incidents, users
  headers.test.ts          14 / 14  — nginx headers, HSTS, CORS
  prompt-guard.test.ts     30 / 30  — injection patterns, false positives
  sockets.test.ts          16 / 16  — chat throttle, monitoring/remediation NS
  stream-tickets.test.ts   10 / 10  — #1112 SSE ticket flow + nginx scrub
  jwt.test.ts              16 / 16  — verifyJwt hardening, env bounds
  infra.test.ts            10 / 16  — TLS, exposure, docker non-root (it.each)
  Total:                  134 / 140

Each file becomes its own vitest worker (fileParallelism: false), so the
vi.resetModules() in JWT_TOKEN_EXPIRY_MINUTES and CORS Zod refinement tests
no longer pollutes sibling describes' Fastify plugin module state.
